@@ -15,7 +15,7 @@ class AccountListExhibit < DisplayCase::Exhibit
     else
       balance = designation_accounts.first.balance.to_i
     end
-    "<div class=\"account_balances tip\" title=\"#{_('May take a few days to update')}\">#{_('Balance: %{balance}') %{balance: @context.number_to_current_currency(balance)}}</div>".html_safe
+    "<div class=\"account_balances tip\" title=\"#{_('May take a few days to update')}\">#{_('Balance: %{balance}').localize % {balance: @context.number_to_current_currency(balance)}}</div>".html_safe
     # <a href=\"#show\" class=\"plain\">Details</a></div>
   end
 
