@@ -6,10 +6,10 @@ class Donation < ActiveRecord::Base
 
   # attr_accessible :donor_account_id, :motivation, :payment_method, :tendered_currency, :donation_date, :amount, :tendered_amount, :currency, :channel, :payment_type
 
-  scope :for, lambda { |designation_account| where(designation_account_id: designation_account.id) }
-  scope :since, lambda { |date| where("donation_date > ?", date) }
+  scope :for, -> (designation_account) { where(designation_account_id: designation_account.id) }
+  scope :since, -> (date) { where("donation_date > ?", date) }
 
-  default_scope order("donation_date desc")
+  default_scope -> { order("donation_date desc") }
 
   after_create :update_totals
   before_validation :set_amount_from_tendered_amount
