@@ -3,7 +3,7 @@ require_dependency 'credential_validator'
 require 'async'
 
 class Person::OrganizationAccount < ActiveRecord::Base
-  extend Person::Account
+  include Person::Account
   include Async
   include Sidekiq::Worker
   sidekiq_options retry: false
@@ -56,7 +56,7 @@ class Person::OrganizationAccount < ActiveRecord::Base
       # we only want to set the last_download date if at least one donation was downloaded
       starting_donation_count = user.donations.count
 
-      update_attributes({downloading: true, locked_at: Time.now}, without_protection: true)
+      update_attributes({downloading: true, locked_at: Time.now})
       date_from = last_download ? (last_download - 2.week) : ''
       organization.api(self).import_all(date_from)
 
@@ -95,8 +95,8 @@ class Person::OrganizationAccount < ActiveRecord::Base
     end
     # If this org account doesn't have any profiles, create a default account list and profile for them
     if user.account_lists.reload.empty? || organization.designation_profiles.where(user_id: person_id).blank?
-      account_list = user.account_lists.create({name: user.to_s, creator_id: user.id}, without_protection: true)
-      organization.designation_profiles.create({name: user.to_s, user_id: user.id, account_list_id: account_list.id}, without_protection: true)
+      account_list = user.account_lists.create({name: user.to_s, creator_id: user.id})
+      organization.designation_profiles.create({name: user.to_s, user_id: user.id, account_list_id: account_list.id})
     end
   end
 
