@@ -4,7 +4,7 @@ require 'rails/all'
 
 if defined?(Bundler)
   # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
+  Bundler.require(:default, Rails.env)
   # If you want your assets lazily compiled in production, use this line
   # Bundler.require(:default, :assets, Rails.env)
 end
@@ -64,6 +64,8 @@ module Mpdx
     #config.assets.precompile += [/(?:\/|\\|\A).*\.css/]
     config.assets.precompile += [/^[a-z0-9]\w+.(css|js)$/]
     config.assets.precompile += ['jquery-ui.css']
+
+    config.active_record.disable_implicit_join_references = true
   end
 end
 
