@@ -1,4 +1,4 @@
 class Version < ActiveRecord::Base
-  attr_accessible :related_object_type, :related_object_id
+  #attr_accessible :related_object_type, :related_object_id
 end
 
