@@ -14,19 +14,19 @@ describe CredentialValidator do
     it "should not add error if the record is missing an org" do
       @record.organization = nil
       @record.valid?
-      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.') % {org: @record.organization})
+      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.').localize % {org: @record.organization})
     end
 
     it "should not add error if the record is missing username" do
       @record.username = nil
       @record.valid?
-      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.') % {org: @record.organization})
+      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.').localize % {org: @record.organization})
     end
 
     it "should add error if the record is missing password" do
       @record.password = nil
       @record.valid?
-      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.') % {org: @record.organization})
+      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.').localize % {org: @record.organization})
     end
 
     it "should add error if the username and password are invalid" do
@@ -34,7 +34,7 @@ describe CredentialValidator do
       @record.organization.stub(:api) {@api}
       @api.stub(:validate_username_and_password) {false}
       @validator.validate(@record)
-      @record.errors.full_messages.should == [_('Your username and password for %{org} are invalid.') % {org: @record.organization}]
+      @record.errors.full_messages.should == [_('Your username and password for %{org} are invalid.').localize % {org: @record.organization}]
     end
 
   end
@@ -48,7 +48,7 @@ describe CredentialValidator do
       @record.username = nil
       @record.password = nil
       @validator.validate(@record)
-      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.') % {org: @record.organization})
+      @record.errors.full_messages.should_not include(_('Your username and password for %{org} are invalid.').localize % {org: @record.organization})
     end
   end
 
